Initial attempt at modernizing for C++11 and Qt 5.
core module:

color/
- Removed color subdirectory, it is now under imageproc and starting
converting it to be a functional style.
- QImageUtils is now gone. It will be back later in an interop module with
  Array2DView<T>.

common/
- Array* and Array*View learned some convenience functions for referenced and spanned bytes. Some minor argument and method renaming.
- ArrayUtils got some function renaming too.
- ArrayUtils learned how to fill an Array2DView and copy().
- BasicTypes.h: removed ubyte/sbyte etc types now that MSVC has cstdint.
  - Also added uint8x2, etc short vector types until vectors are fully
    genericized over type and precision.
- Renamed PingPongBuffers to PingPong.
- ReferenceCountedArray is finally gone.

geometry/
- Added RectangleUtils: utility functions (borrowed from QD3D11) that
  are handy for dealing with rectangles.
  - They will be removed from QD3D11 in the future since this is now core.

imageproc/
- Compositing is now functional and more generic.
- Removed Image* classes: Array2DView is the new hotness.
- Image*'s bilinear sampling functions now live in imageproc/ImageSampling.
- imageproc/Patterns.h and imageproc/Swizzle are now in a functional style.

io/
- Renamed BinaryFileWriter to BinaryFileOutputStream.
- Replaced FileReader with a simple static class File.
  - May convert to functional style later.
- Tweaked io/Portable*IO and io/PNGIO to use new BasicTypes and to support
  slightly more formats.
  - PFMIO now returns a variadic struct: only one field is populated
    depending on the file content.

math/
- BitPacking learned how to swap endianness.
- MathUtils::clampToRange is now generic.
  - Format conversion moved into imageproc/ColorUtils.
- got rid of the poorly thought out UniformFloat/UniformInt classes.
  - STL RNGs will return to fully replace the third-party MersenneTwister
    class used internally by math/Random.
- Got rid of StopWatch now that std::chrono exists.

vecmath/
- Adjusted Quat4d::slerp() to match Quat4f::slerp().

math module:

- Minor tweaks to use cstdint primitives instead of core/common/BasicTypes.

GL module:

- Modernized GLBufferObject for OpenGL 4.4 semantics.
  - Got rid of Usage and instead added Access.
  - Switched to enum class.
  - Got rid of binding and switched to using EXT_direct_state_access.
  - Using Array1DView<T> for mapping.
  - Added a very nice copy() function.
- Got rid of the old GLBufferObjectManager.
  - It was too tied to binding semantics.
  - It will return in some future incarnation as a named hash table to be
    shared amongst objects that use the GPU.
- Aded GLStaticTextureUnitAssignment and GLDynamicTextureUnitAssignment:
  a convenience classes for uniquely mapping a set of textures (and
  samplers) to texture units in a program
  without counting them ourselves.
- GLFramebufferObject is now using EXT_DSA.
- Added enum class wrappers around texture and pixel related GLenums:
  GLImageFormat, GLImageInternalFormat, and GLPixeltype.
  - They're now in separate files instead of inside GLTexture.h.
- Added a simple GLSamplerObject OOP wrapper around OpenGL sampler objects.
- Moved GLShader compilation inside factory constructor.
- Started cleaning up GLTexture*
  - Started using EXT_DSA.
  - Removed dumpTo*: it's simpler to use get() and ArrayUtils instead.
  - Simplified get() and set() functions that use Array*View<T>.
  - Added GLTexture2D::drawNV() a very nice NVIDIA-only extension to
    directly draw textures on screen.
- GLUtilities improvements.
  - Got rid of redundant orthoCamera(): use
    Matrix4f::orthographicProjection() instead.
  - Added clearDepth() and clearDepthNV() for traditional clamped and
    NVIDIA-only unclamped depth buffers.
  - Added get/set viewports, including a "3D" one that has the depth range.
  - Added setDepthRange() and the NVIDIA-only unclamped setDepthRangeNV().
- Added GLVertexArrayObject: a simple wrapper around the modern OpenGL
  vertex array object core feature for vertex specification.
- Removed GLVertexBufferObject: it's been subsumed by the
  modern GLVertexArrayObject and GLBufferObject.

+++ core/src/io/OBJData.cpp
@@ -146,19 +146,19 @@ bool OBJData::save( QString filename )
 	QByteArray ba = filename.toUtf8();
 	FILE* fp = fopen( ba.data(), "w" );
 
-	for( int i = 0; i < m_positions.size(); ++i )
+	for( size_t i = 0; i < m_positions.size(); ++i )
 	{
 		Vector3f v = m_positions[ i ];
 		fprintf( fp, "v %f %f %f\n", v.x, v.y, v.z );
 	}
 
-	for( int i = 0; i < m_textureCoordinates.size(); ++i )
+	for (size_t i = 0; i < m_textureCoordinates.size(); ++i)
 	{
 		Vector2f t = m_textureCoordinates[ i ];
 		fprintf( fp, "vt %f %f\n", t.x, t.y );
 	}
 
-	for( int i = 0; i < m_normals.size(); ++i )
+	for (size_t i = 0; i < m_normals.size(); ++i)
 	{
 		Vector3f n = m_normals[ i ];
 		fprintf( fp, "vn %f %f %f\n", n.x, n.y, n.z );
